TextureManagerの作成を行いました。 DirectXResourceObjectを継承したTextureの作成を行いました。 Texture用にDirectXCommandクラスに関数が追加されました。 GameObjectのコンストラクタにexplicitを付与しました。 その他、軽微な調整、修正が行われています。

+++ Engine/DirectX/DirectXResourceObject/DirectXResourceObject.cpp
@@ -22,7 +22,7 @@ const Microsoft::WRL::ComPtr<ID3D12Resource>& DirectXResourceObject::get_resourc
 	return resource;
 }
 
-[[nodiscard]] Microsoft::WRL::ComPtr<ID3D12Resource> DirectXResourceObject::CreateBufferResource(std::uint32_t sizeInByte) {
+[[nodiscard]] Microsoft::WRL::ComPtr<ID3D12Resource> DirectXResourceObject::CreateBufferResource(std::uint64_t sizeInByte) {
 	Microsoft::WRL::ComPtr<ID3D12Resource> bufferResource;
 	D3D12_HEAP_PROPERTIES uploadHeapProperties{};
 	uploadHeapProperties.Type = D3D12_HEAP_TYPE_UPLOAD; // UploadのHeap
